EXTRACT: create render service, fix unittests DI
