+++ phpunit.xml
@@ -12,12 +12,6 @@
          bootstrap="vendor/autoload.php"
 >
     <testsuites>
-        <testsuite name="Core test suite">
-            <directory suffix="Test.php">modules/core/tests</directory>
-        </testsuite>
-        <testsuite name="Webapp test suite">
-            <directory suffix="Test.php">modules/webapp/tests</directory>
-        </testsuite>
         <testsuite name="Integration tests">
             <directory suffix="Test.php">tests</directory>
         </testsuite>
