+++ tests/ODSReaderTest.php
@@ -4,7 +4,7 @@ class ODSReaderTest extends PHPUnit_Framework_TestCase
 {
     public function testODS()
     {
-        $ods = new \Appizy\Core\ODSReader();
+        $ods = new \Appizy\ODSReader();
         $ods->load(__DIR__ . '/fixtures/demo-appizy.ods');
 
         $tables = $ods->getTables();
@@ -12,7 +12,7 @@ class ODSReaderTest extends PHPUnit_Framework_TestCase
     }
 
     public function testGetRows() {
-        $ods = new \Appizy\Core\ODSReader();
+        $ods = new \Appizy\ODSReader();
         $ods->load(__DIR__ . '/fixtures/demo-appizy.ods');
 
         $tables = $ods->getTables();
@@ -25,7 +25,7 @@ class ODSReaderTest extends PHPUnit_Framework_TestCase
     }
 
     public function testGetCells() {
-        $ods = new \Appizy\Core\ODSReader();
+        $ods = new \Appizy\ODSReader();
         $ods->load(__DIR__ . '/fixtures/demo-appizy.ods');
 
         $tables = $ods->getTables();
