+++ tests/ThemeTest.php
@@ -1,18 +1,18 @@
 <?php
 
-class ThemeTest extends PHPUnit\Framework\TestCase
+class ThemeTest extends PHPUnit_Framework_TestCase
 {
     // TODO: refactor $theme object generation in a setup or before function
 
     public function testGetThemeDirectory(){
-        $theme = new \Appizy\Core\Theme();
+        $theme = new \Appizy\Model\Theme();
         $theme->load(__DIR__ . '/fixtures/theme.info.yml');
 
         $this->assertEquals($theme->getDirectory(), __DIR__ . '/fixtures');
     }
 
     public function testGetTemplateFiles(){
-        $theme = new \Appizy\Core\Theme();
+        $theme = new \Appizy\Model\Theme();
         $theme->load(__DIR__ . '/fixtures/theme.info.yml');
 
         $this->assertEquals(2, count($theme->getTemplateFiles()));
@@ -20,7 +20,7 @@ class ThemeTest extends PHPUnit\Framework\TestCase
 
     public function testGetIncludeFiles()
     {
-        $theme = new \Appizy\Core\Theme();
+        $theme = new \Appizy\Model\Theme();
         $theme->load(__DIR__ . '/fixtures/theme.info.yml');
         $includedFiles = $theme->getIncludedFiles();
