+++ tests/WebAppIntegrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use Appizy\Core\Command\ConvertCommand;
+use Appizy\Command\ConvertCommand;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Tester\CommandTester;
 use Symfony\Component\DomCrawler\Crawler;
